requirements.txt
fastapi
uvicorn
pandas
python-multipart
google-generativeai
supabase
python-dotenv
openpyxl
